moved date functions into utilities
